requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-phonenumber-field==5.2.0
django-recaptcha==2.0.6
django-redis==5.0.0
django-storages==1.11.1
djangorestframework==3.12.4
kombu==5.1.0
ordered-set==4.0.2
phonenumberslite==8.12.36
Pillow==8.4.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
ua-parser==0.10.0
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
